requirements.txt
boto3==1.28.9
branca==0.5.0
elasticsearch==7.13.4
face_recognition==1.3.0
Flask==1.1.4
Flask_API==2.0
Flask_Cors==3.0.8
folium==0.13.0
geopy==2.0.0
numpy==1.24.4
opencv_python==4.6.0.66
python-dotenv==1.0.0
python_dateutil==2.8.1
pytz==2020.4
requests_aws4auth==1.1.1
Shapely==1.7.0
Werkzeug==1.0.1
gunicorn==19.9.0
markupsafe==2.0.1
